creando factory de products

+++ database/migrations/2023_05_08_187912_create_transactions_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained();
             $table->integer('points');
             $table->foreignId('product_id')->constrained();
-            $table->integer('total');
+            $table->float('total');
             $table->timestamps();
         });
     }
